Apply the bits needed to build the errors example.
- Turn off the test module for the saw/crux build; not absolutely sure
why it doesn't work but we don't need or want it in the verification
so it's not worth chasing after (especially since it's probably
related to being on an old SDK version...)

- Also adjust dev_dependencies in Cargo.toml to avoid adding the
testutils feature of the SDK, presumably for related reasons.

These are copied from the other examples we build, and should probably
get applied to the rest of them.