Adjust the saw build for cross_contract.
You need to do both a regular and a saw build of contract_a to build
contract_b, because the contract import mechanism needs to handle the
wasm file produced by the regular build of contract_a. Have saw-build
for contract_a also do a regular build so nobody has to think about
this explicitly.